+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added minimum timeout of 5 seconds for `CohortIdentificationConfigurationSource`
 - 'View Logs' tree view now accessible for CacheProgress objects
 - Console GUI now shows important information (e.g. 'Disabled') in brackets next to items where state is highly important
+- Added new command RunSupportingSql
 
 ### Fixed